+++ README.md
@@ -20,8 +20,8 @@
 
 2) Convert it to numpy array.
 	```bash
-	$ python build_npy_from_bigwig.py test/hg38/ENCFF622DXZ.bigWig
-	$ python build_npy_from_bigwig.py test/hg38/ENCFF074VQD.bigWig
+	$ python bw_to_npy.py test/hg38/ENCFF622DXZ.bigWig
+	$ python bw_to_npy.py test/hg38/ENCFF074VQD.bigWig
 	```
 
 3) Run it. If you score without a variance `.npy` file specified as `--var-npy`, then `msevar` metric will be `0.0`.
@@ -50,12 +50,12 @@ $ python validate.py [YOUR_SUBMISSION_BIGWIG]
 
 2) In order to speed up scoring, convert `TRUTH_BIGWIG` into numpy array/object (binned at `25`). Repeat this for each pair of cell type and assay.
 	```bash
-	$ python build_npy_from_bigwig.py [TRUTH_BIGWIG] --out-npy-prefix [TRUTH_NPY_PREFIX]
+	$ python bw_to_npy.py [TRUTH_BIGWIG] --out-npy-prefix [TRUTH_NPY_PREFIX]
 	```
 
 3) Create a score database.
 	```bash
-	$ python create_db.py [SCORE_DB_FILE]
+	$ python db.py [SCORE_DB_FILE]
 	```
 
 4) For each assay type, build a variance `.npy` file, which calculates a variance for each bin for each chromosome across all cell types. Without this variance file, `msevar` will be `0.0`.
@@ -64,16 +64,15 @@ $ python validate.py [YOUR_SUBMISSION_BIGWIG]
 
 5) Score each submission with bootstrap labels. `--validated` is only for validated submissions binned at `25`. With this flag turned on, `score.py` will skip interpolation of intervals in a bigwig. For ranking, you need to define all metadata for a submission like `--cell [CELL_ID] --assay [ASSAY_OR_MARK_ID] -t [TEAM_ID_INT] -s [SUBMISSION_ID_INT]`. These values will be written to a database file together with bootstrap scores. Repeat this for each submission (one submission per team for each pair of cell type and assay).
 	```bash
-	$ python score.py [YOUR_VALIDATED_SUBMISSION] [TRUTH_NPY] \
-	    --bootstrapped-label-npy [BOOTSTRAP_LABEL_NPY] \
-	    --var-npy var_[ASSAY_OR_MARK_ID].npy
-		--out-db-file [SCORE_DB_FILE] \
+	$ python score.py [YOUR_VALIDATED_SUBMISSION_BIGWIG_OR_NPY] [TRUTH_NPY] \
+	    --var-npy var_[ASSAY_OR_MARK_ID].npy \
+		--db-file [SCORE_DB_FILE] \
 		--cell [CELL_ID] --assay [ASSAY_OR_MARK_ID] \
 		-t [TEAM_ID_INT] -s [SUBMISSION_ID_INT] \
 		--validated
 	```
 
-5) Calculate ranks based on DB file
+6) Calculate ranks based on DB file
 	```bash
 	$ python rank.py [SCORE_DB_FILE]
 	```
